+++ src/components/Icons/Sorting/styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const StyledIcon = styled.svg`
+  path {
+    transition: fill 300ms ease;
+  }
+`;