+++ src/components/MultiSelect/styles.tsx
@@ -1,8 +1,9 @@
-import styled, { css } from "styled-components";
-import { ListWrapper as SingleSelectListWrapper } from "../SingleSelect/styles";
-import { customScrollbar } from "../../mixins/ScrollBar";
+import styled, { css } from 'styled-components';
+import { ListWrapper as SingleSelectListWrapper } from '../SingleSelect/styles';
+import { customScrollbar } from '../../mixins/ScrollBar';
 
-import { isOpenType } from ".";
+import { isOpenType } from '.';
+import { farmerConnectTheme } from '../Theme';
 
 export const Wrapper = styled.div`
   position: relative;
@@ -16,42 +17,30 @@ type HeadingProps = {
 };
 
 export const Heading = styled.div<HeadingProps>`
-  ${(props) => css`
+  ${({ disabled, isOpen }) => css`
     flex: 1;
     display: flex;
     justify-content: space-between;
     align-items: center;
     font-size: 0.875rem;
     line-height: 1.32;
-    padding: 0.625rem 1rem;
-    color: ${props.disabled ? "#B9B9B9" : "#141414"};
-    background-color: ${props.disabled ? "#EAEAEA" : "#F7F6F4"};
+    padding: 0.5rem 0.75rem 0.5rem 1rem;
+    color: ${disabled ? farmerConnectTheme.colors.fc_black_30 : farmerConnectTheme.colors.fc_black_70};
+    background-color: ${disabled ? farmerConnectTheme.colors.fc_black_10 : farmerConnectTheme.colors.fc_beige};
     transition: all 0.05s ease-out;
-    border-radius: ${props.isOpen
-      ? props.isOpen === "left"
-        ? "0.75rem 0 0 0"
-        : "0.75rem 0 0 0.75rem"
-      : "0.75rem 0 0 0.75rem"};
-    color: #5b5b5b;
-    cursor: ${props.disabled ? "default" : "pointer"};
-    box-shadow: ${props.isOpen ? "0.125rem 0.125rem 0 0 #00E394" : "none"};
-    border-right: ${props.isOpen === "left"
-      ? "2px solid #00E394"
-      : "1px solid #b9b9b9"};
+    border-radius: ${isOpen ? (isOpen === 'left' ? '0.75rem 0 0 0' : '0.75rem 0 0 0.75rem') : '0.75rem 0 0 0.75rem'};
+    cursor: ${disabled ? 'default' : 'pointer'};
+    box-shadow: ${isOpen ? `0.125rem 0.125rem 0 0 ${farmerConnectTheme.colors.fc_green}` : 'none'};
+    border-right: ${isOpen === 'left'
+      ? `2px solid ${farmerConnectTheme.colors.fc_green}`
+      : `1px solid ${farmerConnectTheme.colors.fc_black_30}`};
     > svg {
-      fill: ${props.disabled ? "#B9B9B9" : "#5B5B5B"};
-      transform: ${props.isOpen ? "rotate(180deg)" : "none"};
+      transform: ${isOpen ? 'rotate(180deg)' : 'none'};
     }
     & + ${Heading} {
-      border-radius: ${props.isOpen
-        ? props.isOpen === "right"
-          ? "0 0.75rem 0 0"
-          : "0 0.75rem 0.75rem 0"
-        : "0 0.75rem 0.75rem 0"};
+      border-radius: ${isOpen ? (isOpen === 'right' ? '0 0.75rem 0 0' : '0 0.75rem 0.75rem 0') : '0 0.75rem 0.75rem 0'};
       border-right: none;
-      box-shadow: ${props.isOpen === "right"
-        ? "0.125rem 0.125rem 0 0 #00E394"
-        : "none"};
+      box-shadow: ${isOpen === 'right' ? `0.125rem 0.125rem 0 0 ${farmerConnectTheme.colors.fc_green}` : 'none'};
     }
   `}
 `;
@@ -62,19 +51,17 @@ type ContentProps = {
 
 export const Content = styled.div<ContentProps>`
   position: absolute;
-  background-color: #f7f6f4;
-  border-radius: ${(props) =>
-      props.isOpen === "right" ? "0.75rem 0" : "0 0.75rem"}
-    0.75rem 0.75rem;
+  background-color: ${farmerConnectTheme.colors.fc_beige};
+  border-radius: ${({ isOpen }) => (isOpen === 'right' ? '0.75rem 0' : '0 0.75rem')} 0.75rem 0.75rem;
   z-index: 5;
   left: 0;
   right: 0;
   top: 100%;
   display: flex;
   flex-direction: column;
-  height: ${(props) => (props.isOpen ? "20rem" : "0")};
+  height: ${({ isOpen }) => (isOpen ? '20rem' : '0')};
   overflow: hidden;
-  box-shadow: 0.125rem 0.125rem 0 0 #00e394;
+  box-shadow: 0.125rem 0.125rem 0 0 ${farmerConnectTheme.colors.fc_green};
   transition: all 0.05s ease-out;
 `;
 
@@ -93,7 +80,7 @@ export const ProductList = styled.ul`
   flex: 1;
   overflow-y: auto;
 
-  ${customScrollbar({ trackBackgroundColor: '#e7e7e7' })}
+  ${customScrollbar({ trackBackgroundColor: farmerConnectTheme.colors.fc_black_10 })}
 
   > li {
     list-style: none;
@@ -101,7 +88,8 @@ export const ProductList = styled.ul`
     padding: 0.3125rem 1rem;
     cursor: pointer;
   }
+
   li + li {
-    border-top: 1px solid #e7e7e7;
+    border-top: 1px solid ${farmerConnectTheme.colors.fc_black_10};
   }
 `;