+++ src/components/Pagination/styles.ts
@@ -1,5 +1,6 @@
 import { IPaginationProps } from './interfaces';
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
+import Button from '../Button';
 
 export const PageCount = styled.div<any>`
   width: 10rem;
@@ -30,23 +31,30 @@ export const InlinePaginationNumbers = styled.ul<any>`
   }
 `;
 
+const defaultButtonStyle = css`
+  min-width: 0px;
+  font-weight: initial;
+  border-radius: 0.5em;
+  height: 1.5rem;
+
+  > svg {
+    margin: 0 auto;
+  }
+`;
+
+export const PaginationNumberButton = styled(Button)`
+  padding: 0 0.5rem;
+  ${defaultButtonStyle}
+`;
+
+export const PaginationArrowButton = styled(Button)`
+  padding: 0 0.25rem;
+  ${defaultButtonStyle}
+`;
+
 export const PaginationContainer = styled.div<Partial<IPaginationProps>>`
   display: inline-flex;
   gap: 0.5rem;
   width: 100%;
   justify-content: space-between;
-
-  button {
-    padding: 0rem 0.5rem;
-    min-width: 0px;
-    font-weight: initial;
-    border-radius: 0.5em;
-    height: 1.5rem;
-
-    > svg {
-      margin: 0 auto;
-      width: 1rem;
-      height: 1.75rem;
-    }
-  }
 `;