+++ src/components/Select/index.stories.tsx
@@ -12,12 +12,12 @@ export default {
 
 const props = {
   options: [
-    { label: 'Label 1', value: 0 },
-    { label: 'Label 2', value: 1 },
-    { label: 'Label 3', value: 2 },
-    { label: 'Label 4', value: 3 },
-    { label: 'Label 5', value: 4 },
-    { label: 'Label 6', value: 5 },
+    { label: 'Label 1', value: '0' },
+    { label: 'Label 2', value: '1' },
+    { label: 'Label 3', value: '2' },
+    { label: 'Label 4', value: '3' },
+    { label: 'Label 5', value: '4' },
+    { label: 'Label 6', value: '5' },
   ],
 };